Split the deleter job with tasks
